chore: fix failing test
